requirements.txt
absl-py==2.3.1
cffi==2.0.0
contourpy==1.3.0
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.60.2
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
mediapipe==0.10.32
numpy==2.0.2
opencv-contrib-python==4.13.0.90
opencv-python==4.13.0.90
packaging==26.0
pillow==11.3.0
pycparser==2.23
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
sounddevice==0.5.5
zipp==3.23.0

tensorflow
keras
pyyaml
